requirements.txt
# RenderCV Backend Dependencies
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17

# RenderCV engine (this will pull in compatible pydantic)
rendercv[full]>=2.1

# Additional pydantic tools
pydantic-settings>=2.6.0

# YAML handling
pyyaml>=6.0.2

# Async support
aiofiles>=24.1.0

# Image/PDF handling
Pillow>=11.0.0
pdf2image>=1.17.0
cairosvg>=2.7.1
PyMuPDF>=1.26.0

# Health checks
httpx>=0.28.0

# Caching
cachetools>=5.5.0
redis>=5.0.0
